Moved helpers into a subdir added ignore to precommit
